fix TagsCloudVisualization logic
