+++ TagsCloudVisualization/Layouters/TagLayouter.cs
@@ -1,5 +1,6 @@
 using System.Drawing;
 using TagsCloudVisualization.Models;
+using TagsCloudVisualization.Options;
 
 namespace TagsCloudVisualization.Layouters;
 
@@ -11,14 +12,14 @@ public class TagLayouter : ITagLayouter
     private readonly FontFamily _fontFamily;
     private int FontSizeOffset => _maxFontSize - _minFontSize;
     private readonly Graphics _graphics;
-    
+
     public TagLayouter(ICloudLayouter cloudLayouter, TagLayouterOptions options)
     {
         if (options.MinFontSize <= 0)
             throw new ArgumentException($"{nameof(options.MinFontSize)} must be more than zero.");
         if (options.MinFontSize > options.MaxFontSize)
             throw new ArgumentException($"{nameof(options.MinFontSize)} must be less or equal {nameof(options.MaxFontSize)}");
-        
+
         _cloudLayouter = cloudLayouter;
         _minFontSize = options.MinFontSize;
         _maxFontSize = options.MaxFontSize;
@@ -38,9 +39,9 @@ public class TagLayouter : ITagLayouter
         {
             var fontSize = GetFontSize(count, minCount, maxCount);
             yield return new Tag(
-                word, 
-                fontSize, 
-                _fontFamily, 
+                word,
+                fontSize,
+                _fontFamily,
                 _cloudLayouter.PutNextRectangle(GetWordSize(word, fontSize)));
         }
     }
@@ -48,25 +49,6 @@ public class TagLayouter : ITagLayouter
     private int GetFontSize(int count, int minWordCount, int maxWordCount) =>
         _minFontSize + (count - minWordCount) * FontSizeOffset / (maxWordCount - minWordCount);
 
-    private Size GetWordSize(string word, int fontSize) => 
+    private Size GetWordSize(string word, int fontSize) =>
         Size.Ceiling(_graphics.MeasureString(word, new Font(_fontFamily, fontSize)));
-}
-
-public class TagLayouterOptions(int minFontSize, int maxFontSize, string fontName)
-{
-    public int MinFontSize { get; } = minFontSize;
-    public int MaxFontSize { get; } = maxFontSize;
-    public FontFamily FontFamily { get; } = ConvertFontNameToFontFamily(fontName);
-
-    private static FontFamily ConvertFontNameToFontFamily(string fontName)
-    {
-        try
-        {
-            return new FontFamily(fontName);
-        }
-        catch (ArgumentException e)
-        {
-            throw new ArgumentException($"Font name '{fontName}' is invalid", e);
-        }
-    }
 }
