+++ TagsCloudVisualization/TagsCloudImageCreators/ITagsCloudImageCreator.cs
@@ -1,0 +1,6 @@
+namespace TagsCloudVisualization.TagsCloudImageCreators;
+
+public interface ITagsCloudImageCreator
+{
+    void CreateImageWithTags(string pathToText);
+}
