+++ TagsCloudVisualization/TextReaders/TxtReader.cs
@@ -2,13 +2,9 @@ namespace TagsCloudVisualization.TextReaders;
 
 public class TxtReader : ITextReader
 {
-    public bool IsCanRead(string filePath) => 
+    public bool IsCanRead(string filePath) =>
         filePath.EndsWith(".txt");
 
-    public string ReadWords(string filePath)
-    {
-        using var reader = new StreamReader(filePath);
-        
-        return reader.ReadToEnd();
-    }
+    public string ReadWords(string filePath) =>
+        File.ReadAllText(filePath);
 }
